MBF21: ML_BLOCKPLAYERS and ML_BLOCKLANDMONSTERS line flags
Ported from Woof (src/doomdata.h:217, src/p_map.c:419,427): players are
blocked by ML_BLOCKPLAYERS (bit 13), and monsters without MF_FLOAT by
ML_BLOCKLANDMONSTERS (bit 12). Bits 12/13 were free -- nothing above
ML_PASSUSE (1<<9) was in use.

Also read linedef flags through (unsigned short) as Woof does at
src/p_setup.c:388. SHORT() yields a signed short, so a linedef with bit 15
set sign-extended into the int and switched on every high bit, which with
the new flags would have blocked players and land monsters on such lines.

+++ p_setup.c
@@ -402,7 +402,10 @@ void P_LoadLineDefs (int lump)
       line_t *ld = lines+i;
       vertex_t *v1, *v2;
 
-      ld->flags = SHORT(mld->flags);
+      // cast through unsigned short (Woof src/p_setup.c:388): SHORT() yields a
+      // signed short, so a linedef with bit 15 set would sign-extend into the
+      // int and switch on every high bit -- including the mbf21 block flags.
+      ld->flags = (unsigned short)SHORT(mld->flags);
       ld->special = SHORT(mld->special);
       ld->tag = SHORT(mld->tag);
       v1 = ld->v1 = &vertexes[SHORT(mld->v1)];
